+++ dbdaora/hash/_tests/conftest.py
@@ -14,7 +14,13 @@ from dbdaora import (
 
 @pytest.mark.asyncio
 @pytest.fixture
-async def fake_service(mocker, fallback_data_source, fake_hash_repository_cls):
+async def fake_service(
+    mocker,
+    fallback_data_source,
+    fake_hash_repository_cls,
+    has_add_cb,
+    has_delete_cb,
+):
     memory_data_source_factory = partial(
         make_aioredis_data_source,
         'redis://',
@@ -39,6 +45,8 @@ async def fake_service(mocker, fallback_data_source, fake_hash_repository_cls):
         cb_expected_exception=RedisError,
         cb_expected_fallback_exception=KeyError,
         logger=mocker.MagicMock(),
+        has_add_circuit_breaker=has_add_cb,
+        has_delete_circuit_breaker=has_delete_cb,
     )
 
     yield service