+++ dbdaora/hash/_tests/test_integration_service_hash_aioredis_delete.py
@@ -6,6 +6,11 @@ from jsondaora import dataclasses
 from dbdaora.exceptions import EntityNotFoundError
 
 
+@pytest.fixture
+def has_delete_cb():
+    return True
+
+
 @pytest.mark.asyncio
 async def test_should_delete(
     fake_service, serialized_fake_entity, fake_entity