+++ dbdaora/hash/conftest.py
@@ -9,15 +9,30 @@ from dbdaora import (
     CacheType,
     DictFallbackDataSource,
     HashRepository,
-    HashService,
-    build_service,
     make_aioredis_data_source,
+    make_hash_service,
 )
+
+
+@pytest.fixture
+def has_add_cb():
+    return False
+
+
+@pytest.fixture
+def has_delete_cb():
+    return False
 
 
 @pytest.mark.asyncio
 @pytest.fixture
-async def fake_service(mocker, fallback_data_source, fake_hash_repository_cls):
+async def fake_service(
+    mocker,
+    fallback_data_source,
+    fake_hash_repository_cls,
+    has_add_cb,
+    has_delete_cb,
+):
     memory_data_source_factory = partial(
         make_aioredis_data_source,
         'redis://',
@@ -28,8 +43,7 @@ async def fake_service(mocker, fallback_data_source, fake_hash_repository_cls):
     async def fallback_data_source_factory():
         return fallback_data_source
 
-    service = await build_service(
-        HashService,
+    service = await make_hash_service(
         fake_hash_repository_cls,
         memory_data_source_factory,
         fallback_data_source_factory,
@@ -41,6 +55,8 @@ async def fake_service(mocker, fallback_data_source, fake_hash_repository_cls):
         cb_recovery_timeout=10,
         cb_expected_exception=RedisError,
         logger=mocker.MagicMock(),
+        has_add_circuit_breaker=has_add_cb,
+        has_delete_circuit_breaker=has_delete_cb,
     )
 
     yield service